# Added condition to resolve issue in Rhel7.6alt