Based on issue #1406 modified the script to run correctly

Signed-off-by: Shirisha Ganta <shiganta@in.ibm.com>