@shirishaganta 2019-12-11
As you requested above, I changed the script
After Patch:
JOB ID     : 3127a01e7cbd86c98970f4b9794be4b3f25d46c2
JOB LOG    : /home/OpTest/avocado-fvt-wrapper/results/job-2019-12-11T11.41-3127a01/job.log
 (1/3) perf_uprobe.py:PerfUprobe.test_uprobe: PASS (3.98 s)
 (2/3) perf_uprobe.py:PerfUprobe.test_uprobe_return: PASS (2.09 s)
 (3/3) perf_uprobe.py:PerfUprobe.test_uprobe_variable: PASS (1.98 s)
RESULTS    : PASS 3 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 8.74 s
JOB HTML   : /home/OpTest/avocado-fvt-wrapper/results/job-2019-12-11T11.41-3127a01/results.html