@shirishaganta 2019-12-12
> @shirishaganta Does the above condition hold good for Rhel 7.7 and 7.8 as well?

It holds good for both rhel 7.7 and rhel 7.8